+++ actix-limitation/CHANGES.md
@@ -1,7 +1,9 @@
 # Changes
 
 ## Unreleased - 2022-xx-xx
-- Updated `session-session` dependency to `0.7`.
+- `Limiter::builder` now takes an `impl Into<String>`.
+- Removed lifetime from `Builder`.
+- Updated `actix-session` dependency to `0.7`.
 
 
 ## 0.2.0 - 2022-03-22